Move pretty and mlir into separate directories
