+++ arc-script/core/src/codegen/mod.rs
@@ -1,4 +1,3 @@
 pub mod mlir;
 pub mod pretty;
-mod pretty_utils;
 mod printer;
